Macro for all the ids backed by uuids. Those are all the same code.
