+++ crates/types/Cargo.toml
@@ -50,6 +50,7 @@ notify-debouncer-mini = { version = "0.4.1" }
 num-traits = { version = "0.2.17" }
 once_cell = { workspace = true }
 opentelemetry = { workspace = true }
+paste = { workspace = true }
 prost = { workspace = true }
 prost-dto = { workspace = true }
 prost-types = { workspace = true }
